renamed lib to pypackageinstaller
