+++ install.py
@@ -1,4 +1,4 @@
-from lib.install_packages import install
+from pypackageinstaller import install
 
 if __name__ == '__main__': #Only runs automatically if the actual script is run
     install()
