+++ pypackageinstaller/__init__.py
@@ -5,4 +5,4 @@ The above copyright notice and this permission notice shall be included in all
 copies or substantial portions of the Software.
 """
 
-from lib.install_packages import install
+from pypackageinstaller.install_packages import install
